CA-326004: wake polling backend early
When processing a batch wake a (polling) backend twice:
* once as soon as we processed the first request
* at the end of a batch as previously (usually every ring/4)

When polling for request completions we just update our ring index,
which a polling backend should notice but don't trigger an extra evtchn
notification.
When submitting requests we also send a notification on the first
request to wake the backend, by the time it wakes up we might also have
had the chance to put more requests into the ring, which a polling
backend should notice.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>